Split out the router in the http-server example to make the server.js code easier to follow.
